Hello there!
On March 8th, 2021 at 07:58 (UTC) your application (Gitle) used an
access token (with the User-Agent auxilium) as part of a query parameter
to access an endpoint through the GitHub API.

https://api.github.com/user

Please use the Authorization HTTP header instead as using the
`access_token` query parameter is deprecated.

Depending on your API usage, we'll be sending you this email reminder on
a monthly basis.

Visit
https://developer.github.com/changes/2020-02-10-deprecating-auth-through-query-param
for more information about suggested workarounds and removal dates.

Thanks,
The GitHub Team